Module CamomileLibrary

module Private : sig ... end

module ConfigInt : sig ... end

module DefaultConfig = CamomileDefaultConfig

module OOChannel : sig ... end

Individual modules

module UChar : sig ... end

Unicode (ISO-UCS) characters.

module USet : sig ... end

Sets of Unicode characters, implemented as sets of intervals. The signature is mostly same to Set.S in stdlib

module UMap : sig ... end

module UCharTbl : sig ... end

Fast lookup tables for Unicode. Accessible by constant time.

module UnicodeString : sig ... end

Signature for Unicode strings. UText, XString, UTF8, UTF16, UCS4 have matched signatures to UStorage and satisfy the semantics described below. If users want to supply their own Unicode strings, please design the module with the following signature and properties.

module UText : sig ... end

An implementation of Unicode string.

module XString : sig ... end

eXtensible Unicode string. The semantics matches the description of UStorage. The detail may be going to change.

module SubText : sig ... end

module ULine : sig ... end

Line IO

module Locale : sig ... end

Camomile has a locale system similar to Java. A locale is a string with a form as "<LANG>_<COUNTRY>_<MODIFIER>..." where <LANG> is a 2-letter ISO 639 language code, <COUNTRY> is a 2-letter ISO 3166 country code. Some field may not present.

module UTF8 : sig ... end

UTF-8 encoded Unicode strings. The type is normal string.

module UTF16 : sig ... end

module UCS4 : sig ... end

UCS4 encoded string. The type is the bigarray of 32-bit integers. Bigarray.cma or Bigarray.cmxa must be linked when this module is used.

module UPervasives : sig ... end

Functions for toplevel

module URe : sig ... end

Regular expression engine.

module CharEncoding : sig ... end

module UCharInfo : sig ... end

Unicode character informations

module UNF : sig ... end

Unicode normal form (NFD, NFKD, NFC, NFKC) as described in UTR #15

module UCol : sig ... end

Unicode collation algorithm

module CaseMap : sig ... end

Case mappings as defined in Unicode Technical Report #21

module UReStr : sig ... end

Module for a Str-like regular expression syntax. The difference can be summarised as follows.

module StringPrep : sig ... end

All-in-one, configure once at beginning module

module type Type = sig ... end

module Make : functor (Config : ConfigInt.Type) -> sig ... end