Class Configure.out_channel

new out_channel enc outchan creates the output channel object OOChannel.obj_output_channel which receives Unicode characters and outputs them to outchan using the encoding enc.

method close_out : unit -> unit

If close_out cannot output all buffered objects, flush raises Failure

method flush : unit -> unit

If flush cannot output all buffered objects, flush raises Failure

method put : 'a -> unit