Module Interface.Make

Parameters

module Text : UnicodeString.Type

Signature

type text = Text.t

val decode : t -> string -> text

decode enc s converts the string s encoded by the encoding enc to the Unicode text.

val encode : t -> text -> string

encode enc t converts the Unicode text t to the string by the encoding enc.