Class Interface.in_channel

new in_channel enc inchan creates the intput channel object OOChannel.obj_input_channel which reads bytes from inchan and converts them to Unicode characters.

method close_in : unit -> unit

method get : unit -> 'a