Class type OOChannel.obj_input_channel

Generic input channel Have the same interface of Polymorphic input channel of http://www.ocaml-programming.de/rec/IO-Classes.html All channels of Camomile having this interface must confirm the behaviour defined in the recommendation above.

method close_in : unit -> unit

method get : unit -> 'a