Class type OOChannel.obj_output_channel

Generic output channel Have the same interface of Polymorphic output channel of http://www.ocaml-programming.de/rec/IO-Classes.html All channels of Camomile having this interface must confirm the behaviour defined in the recommendation above.

method close_out : unit -> unit

If close_out cannot output all buffered objects, flush raises Failure

method flush : unit -> unit

If flush cannot output all buffered objects, flush raises Failure

method put : 'a -> unit