Module 2-Text.Buf

Character buffers. Similar to Buffer.

type buf

val create : int -> buf

create n creates the buffer. n is used to determine the initial size of the buffer. The meaning of n differs from modules to modules.

val contents : buf -> t

val clear : buf -> unit

val reset : buf -> unit

val add_char : buf -> UChar.t -> unit

val add_string : buf -> t -> unit

val add_buffer : buf -> buf -> unit