Module CamomileLibrary.UReStr

Module for a Str-like regular expression syntax. The difference can be summarised as follows.

Non-ASCII characters can be used if encoded by UTF-8, or using the escape syntax \u<code number as hex digits>.

Each Unicode character is treated as a single character.

Character properties like Lu (UCharInfo.general_category_type), White_Space (UCharInfo.character_property_type), Ogham (UCharInfo.script_type) can be used in character sets. e.g. [{Lu & ID_Start}][{ID_Continue}]* (capitalised identifier), \([{Han}]+\|[{Katakana}]+\)[{Hiragana}]* (Japanese word component). Boolean notations as | (or) :, & (and) - (set subtraction) can be used in {...} notations. Any is used to denote the set of all characters in {...} notations.

module type Interface = sig ... end

module Configure : functor (Config : ConfigInt.Type) -> Interface