Module type CamomileLibrary.Type

module OOChannel : module type of OOChannel

Object Oriented Channel

module USet : module type of USet

Sets of Unicode characters, implemented as sets of intervals. The signature is mostly same to Set.S in stdlib

module UChar : module type of UChar

Unicode (ISO-UCS) characters.

module UMap : module type of UMap

module UCharTbl : module type of UCharTbl

Fast lookup tables for Unicode. Accessible by constant time.

module UnicodeString : module type of UnicodeString

Signature for Unicode strings. UText, XString, UTF8, UTF16, UCS4 have matched signatures to UStorage and satisfy the semantics described below. If users want to supply their own Unicode strings, please design the module with the following signature and properties.

module UText : module type of UText

An implementation of Unicode string.

module XString : module type of XString

eXtensible Unicode string. The semantics matches the description of UStorage. The detail may be going to change.

module SubText : module type of SubText

module ULine : module type of ULine

Line IO

module Locale : module type of Locale

Camomile has a locale system similar to Java. A locale is a string with a form as "<LANG>_<COUNTRY>_<MODIFIER>..." where <LANG> is a 2-letter ISO 639 language code, <COUNTRY> is a 2-letter ISO 3166 country code. Some field may not present.

module CharEncoding : CharEncoding.Interface

module UTF8 : module type of UTF8

UTF-8 encoded Unicode strings. The type is normal string.

module UTF16 : module type of UTF16

module UCS4 : module type of UCS4

UCS4 encoded string. The type is the bigarray of 32-bit integers. Bigarray.cma or Bigarray.cmxa must be linked when this module is used.

module UPervasives : module type of UPervasives

Functions for toplevel

module URe : module type of URe

Regular expression engine.

module UCharInfo : UCharInfo.Type

Character Information

module UNF : sig ... end

module UCol : sig ... end

module CaseMap : sig ... end

module UReStr : UReStr.Interface

module StringPrep : sig ... end