Module type CharEncoding.Type

type text

val decode : t -> string -> text

decode enc s converts the string s encoded by the encoding enc to the Unicode text.

val encode : t -> text -> string

encode enc t converts the Unicode text t to the string by the encoding enc.