Module UCharTbl.Make

Equality and hash are necessary for table generation.

Parameters

module H : Hashtbl.HashedType

Signature

type elt = H.t

type t = elt tbl

val get : elt tbl -> UChar.t -> elt

val of_map : elt -> elt UMap.t -> t

of_map def m creates the table which has the same value to m. The table returns def for the characters for which m is undefined.