Module type UCharTbl.Type

type elt

type t = elt tbl

val get : elt tbl -> UChar.t -> elt

val of_map : elt -> elt UMap.t -> t

of_map def m creates the table which has the same value to m. The table returns def for the characters for which m is undefined.