Module UCol.Make

Parameters

module Text : UnicodeString.Type

Signature

type text = Text.t

type index = Text.index

val compare : ?locale:string -> ?prec:UCol.precision -> ?variable:UCol.variable_option -> text -> text -> int

For locale, see Locale. If locale is omitted, the standard UCA order is used. If prec is omitted, the maximum possible strength is used. If variable is omitted, the default of the locale (usually `Shifted) is used. The meaning of the returned value is similar to Pervasives.compare

val sort_key : ?locale:string -> ?prec:UCol.precision -> ?variable:UCol.variable_option -> text -> string

Binary comparison of sort_key gives the same result as compare. i.e. compare t1 t2 = Pervasives.compare (sort_key t1) (sort_key t2) If the same texts are repeatedly compared, pre-computation of sort_key gives better performance.

val compare_with_key : ?locale:string -> ?prec:UCol.precision -> ?variable:UCol.variable_option -> string -> text -> int

Comparison with the sort key.

val search_with_key : ?locale:string -> ?prec:UCol.precision -> ?variable:UCol.variable_option -> string -> text -> index -> index * index

val search : ?locale:string -> ?prec:UCol.precision -> ?variable:UCol.variable_option -> text -> text -> index -> index * index