Module CamomileLibrary__UnicodeString

Signature for Unicode strings. UText, XString, UTF8, UTF16, UCS4 have matched signatures to UStorage and satisfy the semantics described below. If users want to supply their own Unicode strings, please design the module with the following signature and properties.