Module CamomileLibrary.ULine

Line IO

type separator = [

| `CR

| `LF

| `CRLF

| `NEL

| `LS

| `PS

]

Line separators.

`CR specifies carriage return.

`LF specifies linefeed.

`CRLF specifies the sequence of carriage return and linefeed.

`NEL specifies next line (\u0085).

`LS specifies Unicode line separator (\u2028).

`PS specifies Unicode paragraph separator (\u2029).

class input : separator -> UChar.t CamomileLibrary__.OOChannel.#obj_input_channel -> UChar.t CamomileLibrary__OOChannel.obj_input_channel

new input separator input_obj creates the new input channel object OOChannel.obj_input_channel which reads from input_obj and converts line separators (all of CR, LF, CRLF, NEL, LS, PS) to separator.

class output : separator -> UChar.t CamomileLibrary__.OOChannel.#obj_output_channel -> UChar.t CamomileLibrary__OOChannel.obj_output_channel

new output separator output_obj creates the new output channel object OOChannel.obj_output_channel which receives Unicode characters and converts line separators (all of CR, LF, CRLF, NEL, LS, PS) to separator.

module type Type = sig ... end

module Make (Text : UnicodeString.Type) : Type with type text = Text.t