Class Type.output_line

new output_line ~sp output_obj create the new output channel object OOChannel.obj_output_channel which output each line to output_obj using sp as a line separator. If sp is omitted, linefeed (LF) is used.

method close_out : unit -> unit

If close_out cannot output all buffered objects, flush raises Failure

method flush : unit -> unit

If flush cannot output all buffered objects, flush raises Failure

method put : 'a -> unit