Class Make.nfkc

method close_out : unit -> unit

If close_out cannot output all buffered objects, flush raises Failure

method flush : unit -> unit

If flush cannot output all buffered objects, flush raises Failure

method put : 'a -> unit